Revert "Feature/v3 front end implementations"
