+++ backend/docker-compose.yaml
@@ -1,8 +1,8 @@
 version: "3.5"
 services:
-  mongo:
+  database:
     image: mongo
-    container_name: mongo
+    container_name: db
     ports:
       - 27017:27017
     environment:
@@ -12,17 +12,16 @@ services:
   app:
     build: .
     environment:
-      MONGO_HOST: mongo
-      MONGO_PORT: 27017
-      MONGO_NAME: events
-      MONGO_USER: eventifica
-      MONGO_PASS: 123
-      SECRET: 9b9bed51-fddd-424f-8f4a-fcc77593604f
+      DB_HOST: database
+      DB_PORT: 27017
+      DB_NAME: events
+      DB_USER: eventifica
+      DB_PASS: 123
     ports:
       - 3030:3030
     volumes: 
       - .:/usr/app
     links:
-      - mongo
+      - database
     depends_on:
-      - mongo
+      - database
