+++ backend/package.json
@@ -20,10 +20,8 @@
   },
   "homepage": "https://github.com/ArtyrAlmeida/eventifica#readme",
   "devDependencies": {
-    "@types/bcrypt": "^5.0.0",
     "@types/cors": "^2.8.13",
     "@types/express": "^4.17.17",
-    "@types/jsonwebtoken": "^9.0.2",
     "@types/node": "^18.15.11",
     "@typescript-eslint/eslint-plugin": "^5.58.0",
     "@typescript-eslint/parser": "^5.58.0",
@@ -33,10 +31,8 @@
     "typescript": "^5.0.4"
   },
   "dependencies": {
-    "bcrypt": "^5.1.0",
     "cors": "^2.8.5",
     "express": "^4.18.2",
-    "jsonwebtoken": "^9.0.1",
     "mongoose": "^7.2.1"
   }
 }
