--- backend/src/repository/UserRepository.ts
@@ -1,18 +1,0 @@
-import { UserInterface } from "../interfaces";
-import User from "../models/user";
-
-
-export default class UserRepository {
-    async create(user: UserInterface) {
-        const createdUser = await User.create({...user});
-        
-        return createdUser;
-    }
-
-    async findByEmail(email: string) {
-        const result = await User.findOne({ email });
-
-        return result;
-    }
-
-}
