--- backend/src/routes/router.ts
@@ -1,22 +1,0 @@
-import { Router } from 'express';
-import EventController from '../controllers/EventController';
-import requireAdminAuth from '../middleware/requireAdminAuth';
-import { UserController } from '../controllers/UserController';
-
-const eventsRouter = Router();
-const eventController = new EventController();
-
-eventsRouter.get('/', eventController.find);
-eventsRouter.get('/:id', eventController.findOne);
-eventsRouter.get('/search/:text', eventController.findByText);
-eventsRouter.post('/', requireAdminAuth, eventController.create);
-eventsRouter.patch('/:id', requireAdminAuth, eventController.updateOne);
-eventsRouter.delete('/:id', requireAdminAuth, eventController.deleteOne);
-
-const userRouter = Router();
-const userController = new UserController();
-
-userRouter.post('/login', userController.login);
-userRouter.post('/register', userController.register);
-
-export { eventsRouter, userRouter };
