+++ frontend/package.json
@@ -9,7 +9,6 @@
     "@testing-library/user-event": "^13.5.0",
     "bootstrap": "^5.3.0",
     "react": "^18.2.0",
-    "react-auth-kit": "^2.12.3",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.12.1",
     "react-scripts": "5.0.1",
