+++ frontend/src/components/Home/Home.jsx
@@ -1,3 +1,4 @@
+import { Fragment } from "react"
 import { EventList } from "../Events/EventList"
 import  Map  from '../Map/Map'
